requirements.txt
# SmartClean FastAPI Requirements
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.20
python-multipart==0.0.12
pydantic-settings==2.13.1
psycopg2-binary==2.9.11
asyncpg==0.30.0
redis==5.0.0
alembic==1.14.0
